fix: home pin address return watch and safe
